requirements.txt
pyobjc==11.0
pyobjc-framework-Quartz==11.0
markdown==3.7
PyQt5~=5.15.11
py2app==0.28.8
setuptools==70.3.0
ollama==0.4.7
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0